Added utils.py and renamed test_command to test_commands
