+++ tests/test_commands.py
@@ -1,35 +1,10 @@
 import ast
-import importlib
 import inspect
-import pkgutil
-import types
 from textwrap import dedent
-from typing import List, Dict
-
+from typing import List
 from commands2 import Command, Subsystem
-
+from tests.utils import import_submodules
 from utils.safecommand import SafeMixin
-
-
-def import_submodules(package, recursive=True) -> Dict[str, types.ModuleType]:
-    """ Import all submodules of a module, recursively, including subpackages
-
-    :param package: package (name or actual module)
-    :type package: str | module
-    :rtype: dict[str, types.ModuleType]
-    """
-    if isinstance(package, str):
-        package = importlib.import_module(package)
-    results = {}
-    for loader, name, is_pkg in pkgutil.walk_packages(package.__path__):
-        full_name = package.__name__ + '.' + name
-        try:
-            results[full_name] = importlib.import_module(full_name)
-        except ModuleNotFoundError:
-            continue
-        if recursive and is_pkg:
-            results.update(import_submodules(full_name))
-    return results
 
 
 def get_commands() -> List[Command or None]:
